Cosmetic market purchasing system
The purpose of this project is to present a detailed description of the Cosmetic Market System. It will explain the purpose and features of the system, the interfaces of the system, what the system will do, the constraints under which it must operate and how the system will react to external stimuli. This document is intended for both the stakeholders(customers) and the developers of the system.